Added three tests to DetectorTest
The test currently present in testDetectingSimpleClonesWorks is hard to verify
manually, since the number of tokens cannot be counted easily.

This commit adds two identical files, a.php and b.php, in which the number of
(non-ignored) tokens can be conveniently counted. Since phpcpd (currently)
ignores the opening tag, whitespace, comments and the '=', '+' and ';' the
duplication between these files spans exactly 60 tokens and 20 lines.

Using these two files, three tests cases are added to DetectorTest:
 * One that verifies the duplication is found when minLines = 20 and
   minTokens = 60
 * One that verifies the duplication is ignored when minLines = 20 and
   minTokens = 61
 * One that verifies the duplication is ignored when minLines = 21 and
   minTokens = 60

Since these tests can easily be checked by hand, they can be used to verify
that the detector reports duplications with the correct token and line count,
and ignores duplications that are smaller than the configured thresholds.