Fixed incorrect token count in DefaultStrategy
DefaultStrategy finds duplications using a 'sliding window' of $minTokens
tokens. When it detects that the signature of the tokens in the current window
was seen before it keeps moving its window forward until the signature does not
match an earlier signature anymore (or the end of the file is reached). This
indicates the end of the duplication.

When the end of a duplication is detected, the number of the last duplicated
token was incorrectly calculated. Rather than the start token of the current
window it should be the last token of the previous window, since that window
was duplicated in its entirety. The last token of the duplication is therefore
not $tokenNr, but ($tokenNr - 1) + $minTokens - 1: the start token of the
previous window plus the window length minus 1.

Also the loop that 'moved the window forward' was incorrect; it can stop as
soon as there does not fit an entire window of $minTokens tokens in the file
anymore. In that case there can never be a duplication of $minTokens tokens
at the end of the file anymore.